Fix out of memory by streaming file content at upload
The current upload process was reading in memory the file content several
times:
* one during the API upload to a temporary file
* and in byDefault, byResize and setWH in \Core\Upload

This prevents uploading very large non image files to S3 for instance.
Note that getimagesize used in setWH is only meaningful for images.
Unfortunately, getimagesize is reading the whole file in RAM apparently
When dealing with a non image file (for instance a large zip file).
To solve this issue, we run getimagesize only if the file type is Image.